works on PID parameters communication buffer

+++ src/base/edp/HardwareInterface.h
@@ -109,7 +109,7 @@ public:
 	virtual bool is_impulse_zero(int drive_number) = 0;
 	virtual void reset_position(int i) = 0;
 
-	virtual int set_parameter_now(int drive_number, const int parameter, uint32_t new_value) = 0;
+	virtual int set_parameter_now(int drive_number, const int parameter, uint64_t new_value) = 0;
 };
 
 }
